Support late initialization
