+++ lib/core.js
@@ -7,6 +7,7 @@
  * @callback sandboxFactoryFn
  * @param {string|object} options - a string with a fixture name or a configuration object
  * @param {string} options.fixtureName - a fixture name, the folder of the Hexo testing project
+ * @param {boolean} options.skipInit - set to true if you don't want to initialize Hexo. Default: false
  * @return {HexoContext}
  */
 
@@ -79,10 +80,11 @@ export function createSandbox(Hexo, options) {
     ...options
   }
 
-  return function init (initOptions) {
+  return function sandboxFactoryFn (initOptions) {
     initOptions = initOptions || {}
 
     const name = typeof initOptions == 'string' ? initOptions : initOptions.fixtureName
+    const skipInit = typeof initOptions == 'string' ? false : initOptions.skipInit
 
     const baseFolder = name
       ? path.join(options.fixture_folder, name)
@@ -93,14 +95,37 @@ export function createSandbox(Hexo, options) {
     // Bypass the package.json check and make sure the config file can be loaded
     ctx.env.init = true
 
-    return ctx.init()
-      .then(() => {
-        return Promise.each(options.plugins, pluginPath => {
-          ctx.loadPlugin(pluginPath)
-        })
+    ctx.extend.filter.register('after_init', () => {
+      return Promise.each(options.plugins, pluginPath => {
+        ctx.loadPlugin(pluginPath)
       })
-      .then(() => ctx)
+    }, 0)
+
+    if (skipInit) {
+      return Promise.resolve(ctx)
+    }
+
+    return ctx.init().then(() => ctx)
   }
+}
+
+/**
+ * ```
+ * // CommonJS
+ * const {init} = require('hexo-test-utils')
+ *
+ * // ES2015
+ * import {init} from 'hexo-test-utils'
+ * ```
+ *
+ * Initializes all plugins, generates content from folders. Useful if you need to mock main configuration like `source_dir`
+ * Use with `skipInit` in {@link sandboxFactoryFn}.
+ *
+ * @param {HexoContext} ctx
+ * @return {Promise.<HexoContext>}
+ */
+export function init(ctx) {
+  return ctx.init().then(() => ctx)
 }
 
 /**
