+++ package.json
@@ -45,6 +45,7 @@
     "docdash": "^0.4.0",
     "gh-pages": "^1.1.0",
     "hexo": "^3.6.0",
+    "hexo-renderer-markdown-it": "^3.4.1",
     "jsdoc": "^3.5.5"
   }
 }
