+++ specs/support/fixtures/custom_source_dir/_config.yml
@@ -1,0 +1,1 @@
+permalink: :slug/
